requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
fonttools==4.53.0
gunicorn==22.0.0
pytesseract==0.13.0
Jinja2==3.1.4
matplotlib==3.7.5
numpy==1.24.4
pip==24.0
setuptools==70.3.0
wheel==0.43.0
python-dotenv==1.0.0